requirements.txt
Django==4.2.3
django-allauth==0.55.2
django-debug-toolbar==4.2.0
django-redis==5.3.0
django-environ==0.11.2
django_extensions==3.2.3
djangorestframework==3.14.0
django-filter==23.2
asgiref==3.7.2
Pillow==10.0.0
sqlparse==0.4.4
psycopg2==2.9.7
celery==5.3.4
redis==4.6.0
stripe==6.4.0
flake8==6.1.0
isort==5.12.0
markdown==3.4.4
